+++ lib/objects/voucher_children/unlimited_amount_voucher_with_end_time.dart
@@ -42,7 +42,7 @@ class UnlimitedAmountVoucherWithEndTime
     String? voucherID,
     String? voucherName,
     DateTime? startTime,
-    double? discountValue,
+    int? discountValue,
     int? minimumPurchase,
     int? maxUsagePerPerson,
     DistributionType? distributionType,
