+++ lib/objects/voucher_children/unlimited_percentage_voucher_without_end_time.dart
@@ -41,7 +41,7 @@ class UnlimitedPercentageVoucherWithoutEndTime
     String? voucherID,
     String? voucherName,
     DateTime? startTime,
-    double? discountValue,
+    int? discountValue,
     int? minimumPurchase,
     int? maxUsagePerPerson,
     DistributionType? distributionType,
